Add #commit_changes_to_branch test coverage
There’s a bit too much going on here, but this at least provides a base
level of coverage for refactoring.

We wrap the specs in an `around` block to take us into a tmpdir where
we’ve copied our sample git repo. We then clone that into a working copy
via WithGitRepo and perform our operation (in the spec `subject`). After
we’ve done that, we need to re-clone the repo (to another tmpdir) so
that we can check that we did indeed make the commit and push our
changes.