Allow clone_url to be optional
Now that we can pass a `git` object, we only need to pass `clone_url` if
we're falling back to `default_git`.

Had to create a new `WithGitRepo` in the `assert_raises` block because
using `let(:options)` and the `subject` in the block was raising the
exception somewhere else (I think when `subject` is defined). This is
somewhat temporary anyway, since we'll later extract `default_git` to a
class method constructor.