Add WithGitRepo.with_cloned_repo
Allow `WithGitRepo` to be initialised with a cloned `Git` repo. This is
essentially the same behaviour as initializing without a `git` option,
but will later allow us to remove the `Git` initialisation within
`WithGitRepo`. This will make testing easier and provide more
flexibility – for example, another constructor could create a new
`WithGitRepo` with a working copy.

The implemented tests are not ideal. Ordinarily I'd have done something
along the lines of:

    # WithGitRepo
    def ==(other)
      git == other.git
    end

    # describe WithGitRepo…
    subject { WithGitRepo.with_cloned_repo(clone_url) }
    let(:git) { Git.clone(clone_url) }
    # Assert that the class method constructs a new instance in a
    # specific way:
    it { assert_equal WithGitRepo.new(git: git), subject }

`Git::Base` doesn't define an equality operator though, so it's
really difficult to test in a reliable way.